requirements.txt
#pip install -r requirements.txt
matplotlib
numpy
requests
shapely
osmnx
python-dotenv